FR-#24b6jmz-tooltips some initial work and testing, not quite finished yet

+++ Code/NodeCode/NodeSlots/IntSlot.cs
@@ -59,7 +59,7 @@ namespace MetaMaker
 			
 			if( template.values.ContainsKey( "minStepSize" ) )
 			{
-				template.GetValue( "minStepSize", out double minStepSize );
+				template.GetValue( "minStepSize", out int minStepSize );
 				_field.Step = minStepSize;
 			}
 		}
